requirements.txt
pandas==2.1.4
numpy==1.26.4
matplotlib==3.8.2
seaborn==0.13.1
scikit-learn==1.3.2
tensorflow-cpu==2.15.0
scikeras==0.12.0
streamlit==1.31.0
plotly==5.18.0
joblib==1.3.2
